Add `turboquant_compress` to consolidate the full compression pipeline
The full TurboQuant compression pipeline — normalize a tensor via
`normalize_as_l2_denorm`, quantize the normalized child via
`turboquant_encode_unchecked`, and reattach the stored norms as an
outer `L2Denorm` wrapper — was spelled out three times: in
`TurboQuantScheme::compress`, in `vector_search::compress_turboquant`,
and in the `normalize_and_encode` test helper.

Add `turboquant_compress(input, config, ctx)` alongside the existing
encode helpers and use it in all three call sites. Each caller
collapses to a single line that threads its own config through (the
scheme and `compress_turboquant` keep the `TurboQuantConfig::default()`
choice, and the test helper now takes an explicit config).
