+++ vortex-tensor/src/encodings/turboquant/tests/mod.rs
@@ -33,8 +33,8 @@ use vortex_error::VortexExpect;
 use vortex_error::VortexResult;
 
 use crate::encodings::turboquant::TurboQuantConfig;
+use crate::encodings::turboquant::turboquant_compress;
 use crate::encodings::turboquant::turboquant_encode_unchecked;
-use crate::scalar_fns::l2_denorm::L2Denorm;
 use crate::scalar_fns::l2_denorm::normalize_as_l2_denorm;
 use crate::tests::SESSION;
 use crate::vector::Vector;
@@ -84,18 +84,7 @@ fn normalize_and_encode(
     config: &TurboQuantConfig,
     ctx: &mut vortex_array::ExecutionCtx,
 ) -> VortexResult<ArrayRef> {
-    let l2_denorm = normalize_as_l2_denorm(ext.as_ref().clone(), ctx)?;
-    let normalized = l2_denorm.child_at(0).clone();
-    let norms = l2_denorm.child_at(1).clone();
-    let num_rows = l2_denorm.len();
-
-    let normalized_ext = normalized
-        .as_opt::<Extension>()
-        .vortex_expect("normalized child should be an Extension array");
-    // SAFETY: We just normalized the input via `normalize_as_l2_denorm`.
-    let tq = unsafe { turboquant_encode_unchecked(normalized_ext, config, ctx)? };
-
-    Ok(unsafe { L2Denorm::new_array_unchecked(tq, norms, num_rows) }?.into_array())
+    turboquant_compress(ext.as_ref().clone(), config, ctx)
 }
 
 /// Unwrap an L2Denorm ScalarFnArray into (sorf_child, norms_child).
